tests/requirements.txt
# Test Dependencies for cAgents V7.0

# Core testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-timeout>=2.1.0
pytest-xdist>=3.3.1  # Parallel test execution

# YAML parsing
PyYAML>=6.0.1

# Frontmatter parsing
python-frontmatter>=1.0.0

# JSON schema validation
jsonschema>=4.19.0

# File operations
pathlib2>=2.3.7

# Linting
flake8>=6.1.0
pylint>=2.17.0

# Type checking (optional)
mypy>=1.5.0

# Test reporting
pytest-html>=3.2.0
pytest-json-report>=1.5.0
